Fix fetchExtendedChromInfoFromUCSC("hg38")
A recent change on the UCSC side broke
fetchExtendedChromInfoFromUCSC("hg38"), the workhorse behind
Seqinfo("hg38"). See https://support.bioconductor.org/p/117808/#117831
for the details. With this fix, hg38 is still strictly mapped to
GRCh38 and not to a patched version of it (what assembly hg38 is
based on exactly has become somewhat unclear on the UCSC side).