Fix a regression introduced by recent seqlevelsStyle() improvement
Make sure the improved seqlevelsStyle() setter still allows switching the
style of stuff based on an **arbitrary** GRCh38 patch level to UCSC, not
just the style of stuff based on GRCh38.p12 (GRCh38.p12 is the NCBI
assembly that hg38 is currently mapped to).

For example SNPlocs.Hsapiens.dbSNP144.GRCh38 is based on GRCh38.p2 and we
still want to be able to do:

    library(SNPlocs.Hsapiens.dbSNP144.GRCh38)
    snps <- snpsBySeqname(SNPlocs.Hsapiens.dbSNP144.GRCh38, "MT")
    seqlevelsStyle(snps) <- "UCSC"

The regression (introduced by commit 11119c2f) was breaking the above use
case.