Implement getChromInfoFromUCSC() "offline mode" for a selection of genomes
This "offline mode" is supported only for a selection of UCSC genomes and
when getChromInfoFromUCSC() is called with 'assembled.molecules.only=TRUE'.
It works thanks to a collection of tab-delimited files stored in the package
that contain the "assembled molecules info" for the supported genomes.
See README.TXT in GenomeInfoDb/inst/extdata/assembled_molecules_info/UCSC/
for more information.

Calling getChromInfoFromUCSC() with 'assembled.molecules.only=FALSE' (the
default), or with 'recache=TRUE', will trigger retrieval of the chromosome
info from UCSC and issue a warning if this info no longer matches the
"assembled molecules info" stored in the package.