UCSC genome hg38 is now based on GRCh38.p14 instead of GRCh38.p13
However, it's worth noting that hg38 has 2 extra sequences that do NOT
belong to GRCh38.p14: chr11_KQ759759v1_fix and chr22_KQ759762v1_fix.
This is a serious potential source of confusion. Here is why:

FIX PATCH sequences HG107_PATCH and HG1311_PATCH in GRCh38.p13 have been
replaced with FIX PATCH sequences HG107_HG2565_PATCH and HG1311_HG2539_PATCH
in GRCh38.p14. When they come up with a new patch level for GRCh38, the
Genome Reference Consortium usually only **adds** new sequences w.r.t. the
previous patch level, so this is admittedly quite unusual (and not something
that the UCSC folks can control). Anyways, when UCSC sneakily modified their
hg38 genome towards the end of January 2023 to base it on GRCh38.p14 instead
of GRCh38.p13, they did the following:

- They kept the old HG107_PATCH and HG1311_PATCH sequences, even
  though these sequences do not belong to GRCh38.p14. These are
  named chr11_KQ759759v1_fix and chr22_KQ759762v1_fix in hg38.
  I suppose the reason they did this is to not loose the anotations
  attached to these sequences like genes, transcripts, etc...

- They also included replacements sequences HG107_HG2565_PATCH
  and HG1311_HG2539_PATCH. These are named chr11_KQ759759v2_fix
  and chr22_KQ759762v2_fix in hg38.

This means that hg38 has 2 extra sequences w.r.t. GRCh38.p14 (711 sequences
in hg38 vs 709 sequences in GRCh38.p14). But it also means that, even if
hg38 is mostly based on GRCh38.p14, it is still partly based on GRCh38.p13.
Concretely this means that the UCSC genome browser for hg38 can still report
genes or transcripts that are located on GRCh38.p13 but not on GRCh38.p14,
despite the fact that hg38 is officially based on the latter.

Just something to be aware of!