GenomeInfoDb 1.45.11: Fix getChromInfoFromNCBI("Pan_troglodytes-2.1.4")
In September 2025, someone modified the NCBI assembly report for
Pan_troglodytes-2.1.4 (GCF_000001515.5). Among other things, they
added a duplicated entry for chromosome 21 (RefSeqAccn NC_006488.2),
and this addition broke getChromInfoFromNCBI("Pan_troglodytes-2.1.4").
This commit repairs getChromInfoFromNCBI() by dropping the NC_006488.2
row. The idea is to keep the original chromosome 21 because it has a
length that is equal to the length of chr21 in UCSC genome panTro4.

Also, it seems that, as part of the September 2025 changes to
GCF_000001515.5, some sequences were renamed, with the consequence
that test_switching_Seqinfo_to_RefSeq_style() now fails on panTro4.
So this commit also excludes panTro4 from this test.

Finally, downloading GCF_000001515.5_Pan_troglodytes-2.1.4_genomic.fna.gz
with BSgenomeForge::downloadGenomicSequencesFromNCBI("GCF_000001515.5")
reveals that the RefSeq accessions in the FASTA file differ from those
in the NCBI assembly report for thousands of sequences!

So overall the situation with GCF_000001515.5 is not looking good to say
the least.